Standardize 'int Time' to an unsigned long
(This is a big deal on linux, where the high bit is often set!)

+++ Src/DasherCore/AlternatingDirectMode.cpp
@@ -88,7 +88,7 @@ bool CAlternatingDirectMode::DecorateView(CDasherView *pView, CDasherInput *pInp
 }
 
 
-void CAlternatingDirectMode::DirectKeyDown(int iTime, int iId, CDasherView *pView, CDasherModel *pModel) {
+void CAlternatingDirectMode::DirectKeyDown(unsigned long iTime, int iId, CDasherView *pView, CDasherModel *pModel) {
   int iTargetBox;
   switch(iId) {
     case 2:
