+++ Src/DasherCore/AlternatingDirectMode.h
@@ -34,7 +34,7 @@ namespace Dasher {
   void SetupBoxes();
 
  private:
-  void DirectKeyDown(int iTime, int iId, CDasherView *pView, CDasherModel *pModel);
+  void DirectKeyDown(unsigned long iTime, int iId, CDasherView *pView, CDasherModel *pModel);
 
   int m_iLastBox;
 };
