+++ Src/DasherCore/CircleStartHandler.h
@@ -13,7 +13,7 @@ public:
   CCircleStartHandler(CDefaultFilter *pCreator);
   ~CCircleStartHandler();
   virtual bool DecorateView(CDasherView *pView);
-  virtual void Timer(int iTime, dasherint iX, dasherint iY, CDasherView *pView);
+  virtual void Timer(unsigned long iTime, dasherint iX, dasherint iY, CDasherView *pView);
   virtual void HandleEvent(int iParameter);
   virtual void HandleEvent(CDasherView *pView);
